+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- Composite delegate
+
+
 ## 0.2.0 - 2017-08-29
 
 ### Changed